Add options to Album and Artist and change __new__
Add _use_name_for_repr class variable to Album and Artist. If set to
True, the repr() of an Album or Artist object will show its name. This
is intended for interpreter use only.

Add `clear()` class method to Album and Artist as shorthand for
`__new__.clear()`.

Make __new__ Cache for Album and Artist ignore automatic `cls` arg,
and store IDs as strings instead of tuples of strings.